+++ Core/Init.lua
@@ -2,7 +2,7 @@ local addonName, ns = ...
 
 _G.EQObjectiveTracker = ns
 ns.NAME    = addonName
-ns.VERSION = "1.3.0"
+ns.VERSION = "1.3.1"
 
 ns.modules     = {}
 ns.moduleOrder = {}